Extracted more Classes out of NugetHelper
