+++ Assets/NuGet/Editor/Models/UnityVersion.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2db89b7c77bc4708b75703db5ee2fce5
+timeCreated: 1624023247
