+++ Assets/NuGet/Editor/Nuget.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 07d5051b7b3249389fc2894bd9ba54de
+timeCreated: 1624023985
