+++ Assets/NuGet/Editor/Nuget/NugetCLI.cs
@@ -1,0 +1,147 @@
+﻿using System.Diagnostics;
+using System.IO;
+using System.Text;
+using NuGet.Editor.Models;
+using NuGet.Editor.Util;
+using UnityEngine;
+using Debug = UnityEngine.Debug;
+
+namespace NuGet.Editor.Nuget
+{
+    public class NugetCLI
+    {
+        /// <summary>
+        /// The amount of time, in milliseconds, before the nuget.exe process times out and is killed.
+        /// </summary>
+        private readonly int TimeOut;
+
+        public NugetCLI(int timeOut = 60000)
+        {
+            TimeOut = timeOut;
+        }
+
+        /// <summary>
+        /// Runs nuget.exe using the given arguments.
+        /// </summary>
+        /// <param name="arguments">The arguments to run nuget.exe with.</param>
+        /// <param name="logOuput">True to output debug information to the Unity console.  Defaults to true.</param>
+        /// <returns>The string of text that was output from nuget.exe following its execution.</returns>
+        private void RunNugetProcess(string arguments, bool logOuput = true)
+        {
+            // Try to find any nuget.exe in the package tools installation location
+            string toolsPackagesFolder = Path.Combine(Application.dataPath, "../Packages");
+
+            // create the folder to prevent an exception when getting the files
+            Directory.CreateDirectory(toolsPackagesFolder);
+
+            string[] files = Directory.GetFiles(toolsPackagesFolder, "nuget.exe", SearchOption.AllDirectories);
+            if (files.Length > 1)
+            {
+                Debug.LogWarningFormat("More than one nuget.exe found. Using first one.");
+            }
+            else if (files.Length < 1)
+            {
+                Debug.LogWarningFormat("No nuget.exe found! Attemping to install the NuGet.CommandLine package.");
+                NugetHelper.InstallIdentifier(new NugetPackageIdentifier("NuGet.CommandLine", "2.8.6"));
+                files = Directory.GetFiles(toolsPackagesFolder, "nuget.exe", SearchOption.AllDirectories);
+                if (files.Length < 1)
+                {
+                    Debug.LogErrorFormat("nuget.exe still not found. Quiting...");
+                    return;
+                }
+            }
+
+            NugetHelper.LogVerbose("Running: {0} \nArgs: {1}", files[0], arguments);
+
+            string fileName = string.Empty;
+            string commandLine = string.Empty;
+
+#if UNITY_EDITOR_OSX
+            // ATTENTION: you must install mono running on your mac, we use this mono to run `nuget.exe`
+            fileName = "/Library/Frameworks/Mono.framework/Versions/Current/Commands/mono";
+            commandLine = " " + files[0] + " " + arguments;
+            LogVerbose("command: " + commandLine);
+#else
+            fileName = files[0];
+            commandLine = arguments;
+#endif
+            Process process = Process.Start(
+                new ProcessStartInfo(fileName, commandLine)
+                {
+                    RedirectStandardError = true,
+                    RedirectStandardOutput = true,
+                    UseShellExecute = false,
+                    CreateNoWindow = true,
+                    // WorkingDirectory = Path.GettargetFramework(files[0]),
+
+                    // http://stackoverflow.com/questions/16803748/how-to-decode-cmd-output-correctly
+                    // Default = 65533, ASCII = ?, Unicode = nothing works at all, UTF-8 = 65533, UTF-7 = 242 = WORKS!, UTF-32 = nothing works at all
+                    StandardOutputEncoding = Encoding.GetEncoding(850)
+                });
+
+            if (!process.WaitForExit(TimeOut))
+            {
+                Debug.LogWarning("NuGet took too long to finish.  Killing operation.");
+                process.Kill();
+            }
+
+            string error = process.StandardError.ReadToEnd();
+            if (!string.IsNullOrEmpty(error))
+            {
+                Debug.LogError(error);
+            }
+
+            string output = process.StandardOutput.ReadToEnd();
+            if (logOuput && !string.IsNullOrEmpty(output))
+            {
+                Debug.Log(output);
+            }
+        }
+
+        /// <summary>
+        /// Calls "nuget.exe pack" to create a .nupkg file based on the given .nuspec file.
+        /// </summary>
+        /// <param name="nuspecFilePath">The full filepath to the .nuspec file to use.</param>
+        public void Pack(string nuspecFilePath)
+        {
+            if (!Directory.Exists(NugetHelper.PackOutputDirectory))
+            {
+                Directory.CreateDirectory(NugetHelper.PackOutputDirectory);
+            }
+
+            // Use -NoDefaultExcludes to allow files and folders that start with a . to be packed into the package
+            // This is done because if you want a file/folder in a Unity project, but you want Unity to ignore it, it must start with a .
+            // This is especially useful for .cs and .js files that you don't want Unity to compile as game scripts
+            string arguments = string.Format("pack \"{0}\" -OutputDirectory \"{1}\" -NoDefaultExcludes", nuspecFilePath, NugetHelper.PackOutputDirectory);
+
+            RunNugetProcess(arguments);
+        }
+
+        /// <summary>
+        /// Calls "nuget.exe push" to push a .nupkf file to the the server location defined in the NuGet.config file.
+        /// Note: This differs slightly from NuGet's Push command by automatically calling Pack if the .nupkg doesn't already exist.
+        /// </summary>
+        /// <param name="nuspec">The NuspecFile which defines the package to push.  Only the ID and Version are used.</param>
+        /// <param name="nuspecFilePath">The full filepath to the .nuspec file to use.  This is required by NuGet's Push command.</param>
+        /// /// <param name="apiKey">The API key to use when pushing a package to the server.  This is optional.</param>
+        public void Push(NuspecFile nuspec, string nuspecFilePath, string apiKey = "")
+        {
+            string packagePath = Path.Combine(NugetHelper.PackOutputDirectory, string.Format("{0}.{1}.nupkg", nuspec.Id, nuspec.Version));
+            if (!File.Exists(packagePath))
+            {
+                NugetHelper.LogVerbose("Attempting to Pack.");
+                Pack(nuspecFilePath);
+
+                if (!File.Exists(packagePath))
+                {
+                    Debug.LogErrorFormat("NuGet package not found: {0}", packagePath);
+                    return;
+                }
+            }
+
+            string arguments = string.Format("push \"{0}\" {1} -configfile \"{2}\"", packagePath, apiKey, NugetHelper.NugetConfigFilePath);
+
+            RunNugetProcess(arguments);
+        }
+    }
+}
