+++ Assets/NuGet/Editor/Nuget/NugetCLI.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 560a5f493e624064994c2385d388643c
+timeCreated: 1624024016
