+++ Assets/NuGet/Editor/Services/INugetService.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6f30221a93424d54b3d189047ba8bf2c
+timeCreated: 1624033086
