+++ Assets/NuGet/Editor/Services/NugetService.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0111452f4f05468f9d9af32c8f8e786f
+timeCreated: 1624031004
