+++ Assets/NuGet/Editor/Util/CredentialProviderExitCode.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8c3ef7a6d37b43df999010ed017a8e65
+timeCreated: 1624025849
