+++ Assets/NuGet/Editor/Util/CredentialProviderResponse.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 12e2c84a3aff4d028a8d9a2e4af06663
+timeCreated: 1624025759
