+++ Assets/NuGet/Editor/Util/IFileHelper.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b175e1358cec429c90022af20bb8fb0c
+timeCreated: 1624025302
